add quick options menu
