+++ libs/client/ui/src/lib/components/auth/user-menu/index.ts
@@ -2,10 +2,12 @@ import { UserMenuComponent } from './user-menu.component';
 import { FriendsListComponent } from './friends-list/friends-list.component';
 import { MessagesFeedComponent } from './messages-feed/messages-feed.component';
 import { NotificationsListComponent } from './notifications-list/notifications-list.component';
+import { QuickOptionsComponent } from './quick-options/quick-options.component';
 
 export const UserMenuComponents = [
     UserMenuComponent,
     FriendsListComponent,
     MessagesFeedComponent,
     NotificationsListComponent,
+    QuickOptionsComponent,
 ];
