+++ libs/client/ui/src/lib/components/auth/user-menu/notifications-list/notifications-list.component.html
@@ -1,1 +1,14 @@
-notifications
+<div class="m-2">
+    <div class="flex items-center">
+        <div class="flex-1">
+            {{ notifications.numSelected }} selected
+        </div>
+        <button>
+            <rmx-icon name="check-double-line"></rmx-icon>
+            <span>Mark as read</span>
+        </button>
+    </div>
+    <div>
+
+    </div>
+</div>
