+++ libs/client/ui/src/lib/components/auth/user-menu/notifications-list/notifications-list.component.ts
@@ -1,7 +1,10 @@
 import { Component } from '@angular/core';
+import { NotificationsQuery } from '@dragonfish/client/repository/notifications';
 
 @Component({
     selector: 'dragonfish-notifications-list',
     templateUrl: './notifications-list.component.html',
 })
-export class NotificationsListComponent {}
+export class NotificationsListComponent {
+    constructor(public notifications: NotificationsQuery) {}
+}
