+++ libs/client/ui/src/lib/components/auth/user-menu/quick-options/quick-options.component.html
@@ -1,0 +1,24 @@
+<div class="flex flex-col mx-4 mt-4">
+    <div
+        class="button border-gray-600 dark:border-white"
+        [routerLink]="['/profile', pseudQuery.currentId, pseudQuery.current.userTag]"
+    >
+        <rmx-icon name="user-line"></rmx-icon>
+        <span>View Profile</span>
+    </div>
+    <div class="w-full border-b border-gray-600 dark:border-white my-4"></div>
+    <div
+        class="button border-gray-600 dark:border-white"
+        [routerLink]="['/registration/select-pseud']"
+    >
+        <rmx-icon name="group-2-line"></rmx-icon>
+        <span>Switch Profile</span>
+    </div>
+    <div
+        class="button border-gray-600 dark:border-white"
+        (click)="logout()"
+    >
+        <rmx-icon name="logout-circle-line"></rmx-icon>
+        <span>Log Out</span>
+    </div>
+</div>
