+++ libs/client/ui/src/lib/components/auth/user-menu/quick-options/quick-options.component.scss
@@ -1,0 +1,10 @@
+div.button {
+    @apply flex items-center p-4 my-1.5 border rounded-md cursor-pointer transition transform;
+    rmx-icon {
+        @apply mr-2.5;
+    }
+    &:hover {
+        @apply text-white scale-105;
+        background: var(--accent);
+    }
+}
